feat: 그리드 형식 질문 고려를 위한 Question 객체에서 Item 객체 분리
